+++ paywall/src/__tests__/components/helpers/BalanceProvider.test.js
@@ -4,15 +4,35 @@ import * as rtl from 'react-testing-library'
 import { BalanceProvider } from '../../../components/helpers/BalanceProvider'
 
 describe('BalanceProvider Component', () => {
-  function renderIt({ amount, conversion = { USD: 195.99 }, render }) {
+  function renderIt({
+    amount,
+    conversion = { USD: 195.99 },
+    convertCurrency = true,
+    render,
+  }) {
     return rtl.render(
       <BalanceProvider
         amount={amount}
         conversion={conversion}
         render={render}
+        convertCurrency={convertCurrency}
       />
     )
   }
+
+  it('does not convert if convertCurrency is false', () => {
+    expect.assertions(2)
+    renderIt({
+      amount: null,
+      conversion: {},
+      convertCurrency: false,
+      render: (ethValue, fiatValue) => {
+        expect(ethValue).toEqual(' - ')
+        expect(fiatValue).toEqual(' - ')
+      },
+    })
+  })
+
   it('renders with - when amount is null (probably unset)', () => {
     expect.assertions(2)
     renderIt({
@@ -126,7 +146,7 @@ describe('BalanceProvider Component', () => {
       renderIt({
         amount,
         render: ethValue => {
-          expect(ethValue).toEqual('1.99')
+          expect(ethValue).toEqual('2.00')
         },
       })
     })
@@ -155,7 +175,7 @@ describe('BalanceProvider Component', () => {
       renderIt({
         amount,
         render: (ethValue, fiatValue) => {
-          expect(ethValue).toEqual('2000.00')
+          expect(ethValue).toEqual('2,000')
           expect(fiatValue).toEqual('392k')
         },
       })
@@ -170,7 +190,7 @@ describe('BalanceProvider Component', () => {
       renderIt({
         amount,
         render: (ethValue, fiatValue) => {
-          expect(ethValue).toEqual('20000.00')
+          expect(ethValue).toEqual('20,000')
           expect(fiatValue).toEqual('3.9m')
         },
       })
@@ -185,7 +205,7 @@ describe('BalanceProvider Component', () => {
       renderIt({
         amount,
         render: (ethValue, fiatValue) => {
-          expect(ethValue).toEqual('20000000.00')
+          expect(ethValue).toEqual('20m')
           expect(fiatValue).toEqual('3.9b')
         },
       })