+++ unlock-js/src/__tests__/unlockProvider.test.js
@@ -25,48 +25,69 @@ const key = {
 }
 const publicKey = '0x88a5C2d9919e46F883EB62F7b8Dd9d0CC45bc290'
 const password = 'foo'
-
-const userData = {
-  types: {
-    EIP712Domain: [
-      { name: 'name', type: 'string' },
-      { name: 'version', type: 'string' },
-      { name: 'chainId', type: 'uint256' },
-      { name: 'verifyingContract', type: 'address' },
-      { name: 'salt', type: 'bytes32' },
-    ],
-    User: [
-      { name: 'emailAddress', type: 'string' },
-      { name: 'publicKey', type: 'address' },
-      { name: 'passwordEncryptedPrivateKey', type: 'string' },
-    ],
-  },
-  domain: {
-    name: 'Unlock',
-    version: '1',
-  },
-  primaryType: 'User',
-  message: {
-    user: {
-      emailAddress: 'geoff@bitconnect.gov',
-      publicKey: publicKey,
-      passwordEncryptedPrivateKey: key,
-    },
-  },
-}
+const emailAddress = 'geoff@bitconnect.gov'
 
 describe('Unlock Provider', () => {
   let provider
   beforeEach(async () => {
     const readOnlyProvider = 'http://localhost:8545'
     provider = new UnlockProvider({ readOnlyProvider })
-    await provider.connect({ key, password })
+    await provider.connect({ key, password, emailAddress })
   })
 
   describe('object properties', () => {
     it('should have a property `isUnlock` that is set to `true`', () => {
       expect.assertions(1)
       expect(provider.isUnlock).toBeTruthy()
+    })
+
+    it('should have a property `wallet` that is set to an ethers wallet', () => {
+      expect.assertions(1)
+      expect(provider.wallet).toEqual(
+        expect.objectContaining({
+          address: publicKey,
+        })
+      )
+    })
+
+    it('should have a property `emailAddress` that is set to the provided email address', () => {
+      expect.assertions(1)
+      expect(provider.emailAddress).toEqual(emailAddress)
+    })
+
+    it('should have a property `passwordEncryptedPrivateKey` that is set to the provided key', () => {
+      expect.assertions(1)
+      expect(provider.passwordEncryptedPrivateKey).toEqual(key)
+    })
+  })
+
+  describe('signing data', () => {
+    describe('signUserData', () => {
+      it('should sign an object with all fields passed', () => {
+        expect.assertions(1)
+        const input = {
+          emailAddress,
+          publicKey,
+          passwordEncryptedPrivateKey: key,
+        }
+
+        const output = provider.signUserData(input)
+        // sigutil seems to downcase things
+        expect(sigUtil.recoverTypedSignature(output)).toEqual(
+          publicKey.toLowerCase()
+        )
+      })
+
+      it('should also sign an object with default values when not everything is passed', () => {
+        expect.assertions(1)
+        const input = {}
+
+        const output = provider.signUserData(input)
+        // sigutil seems to downcase things
+        expect(sigUtil.recoverTypedSignature(output)).toEqual(
+          publicKey.toLowerCase()
+        )
+      })
     })
   })
 
@@ -76,18 +97,6 @@ describe('Unlock Provider', () => {
       const accounts = await provider.send('eth_accounts')
       expect(accounts).toHaveLength(1)
       expect(accounts[0]).toEqual(publicKey)
-    })
-
-    it('should respond to eth_signTypedData with a valid signature', async () => {
-      expect.assertions(1)
-      const sig = await provider.send('eth_signTypedData', [
-        'account',
-        { data: userData },
-      ])
-      // sigutil seems to downcase things
-      expect(sigUtil.recoverTypedSignature({ data: userData, sig })).toEqual(
-        publicKey.toLowerCase()
-      )
     })
   })
 })