+++ unlock-js/src/v11/purchaseKey.js
@@ -1,7 +1,8 @@
-import Web3Utils from '../utils'
+import utils from '../utils'
 import { GAS_AMOUNTS } from '../constants'
 import TransactionTypes from '../transactionTypes'
 import Errors from '../errors'
+import { approveTransfer } from '../erc20'
 
 /**
  * Purchase a key to a lock by account.
@@ -15,14 +16,40 @@ import Errors from '../errors'
  * @param {string} data
  * @param {string} account
  */
-export default async function(lockAddress, owner, keyPrice) {
+export default async function(
+  lockAddress,
+  owner,
+  keyPrice,
+  account,
+  data,
+  erc20Address
+) {
   const lockContract = await this.getLockContract(lockAddress)
+  // TODO, use the actual decimals from the contract. We assume 18 here because it is the most frequent default.
+
+  const actualAmount = utils.toDecimal(keyPrice, 18)
+
+  const purchaseForOptions = {
+    gasLimit: GAS_AMOUNTS.purchaseFor, // overrides default value for transaction gas price
+  }
+
+  if (erc20Address) {
+    await approveTransfer(
+      erc20Address,
+      lockAddress,
+      actualAmount,
+      this.provider
+    )
+  } else {
+    purchaseForOptions.value = actualAmount
+  }
+
   let transactionPromise
   try {
-    transactionPromise = lockContract['purchaseFor(address)'](owner, {
-      gasLimit: GAS_AMOUNTS.purchaseFor, // overrides default value for transaction gas price
-      value: Web3Utils.toWei(keyPrice, 'ether'), // overrides default value
-    })
+    transactionPromise = lockContract['purchaseFor(address)'](
+      owner,
+      purchaseForOptions
+    )
     const ret = await this._handleMethodCall(
       transactionPromise,
       TransactionTypes.KEY_PURCHASE